# planning-poker/libs/api-client/src/generated/model/addIssueRequest.ts
/**
 * Planning Poker Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request body for adding a new issue
 */
export interface AddIssueRequest { 
    /**
     * Issue title
     */
    title: string;
    /**
     * Optional issue link
     */
    link?: string;
}