# planning-poker/libs/api-client/src/generated/model/finishIssueRequest.ts
/**
 * Planning Poker Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request body for finishing an issue voting round
 */
export interface FinishIssueRequest { 
    /**
     * Average after voting
     */
    average?: string;
    /**
     * Manuale agreed Average after voting
     */
    agreedAverage?: string;
}