# planning-poker/libs/api-client/src/generated/model/issue.ts
/**
 * Planning Poker Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * An issue to be estimated in planning poker
 */
export interface Issue { 
    /**
     * Unique issue identifier
     */
    id: string;
    /**
     * Issue title
     */
    title: string;
    /**
     * Optional issue/ticket link
     */
    link?: string;
    /**
     * Current status of the issue
     */
    status: Issue.StatusEnum;
    /**
     * Average after voting
     */
    average?: string;
    /**
     * Manuale agreed Average after voting
     */
    agreedAverage?: string;
}
export namespace Issue {
    export type StatusEnum = 'pending' | 'voting' | 'done';
    export const StatusEnum = {
        pending: 'pending' as StatusEnum,
        voting: 'voting' as StatusEnum,
        done: 'done' as StatusEnum
    };
}