# planning-poker/libs/api-client/src/generated/model/roomUpdate.ts
/**
 * Planning Poker Backend API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Issue } from './issue';
import { User }  from './user';
import { Room }  from './room';

/**
 * SSE event payload with current room state and users
 */
export interface RoomUpdate {
    room: Room;
    /**
     * All users currently in the room
     */
    users: Array<User>;
    /**
     * All issues in the room
     */
    issues: Array<Issue>;
}